back/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-money==1.1
django==3.0.8
djangorestframework-gis==0.15
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
gdal==3.1.0
idna==2.10
importlib-metadata==1.7.0 ; python_version < '3.8'
markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.5
py-moneyed==0.8.0
pyjwt==1.7.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.24.0
sqlparse==0.3.1
unidecode==1.1.1
urllib3==1.25.9
whitenoise==5.2.0
zipp==3.1.0
